Moved all components that were touched by redux into a containers folder
