+++ src/Containers/Form/Form.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import { login, signUp, favouritesList, hasError } from '../../actions'
-import MoviesContainer from '../MoviesContainer/MoviesContainer'
+import MoviesContainer from '../../Components/MoviesContainer/MoviesContainer'
 import { Route, Link } from 'react-router-dom';
 import { loginVerification, signUpVerification, getFavourites } from '../../apiCalls'
